+++ App/controllers/__init__.py
@@ -1,2 +1,4 @@
 from .user import *
 from .auth import *
+from .routine import *
+from .workout import *
