+++ App/controllers/workout.py
@@ -1,0 +1,8 @@
+def get_all_workouts():
+    return Workout.query.all()
+
+def get_workout_by_id(id):
+    return workout.query.filter_by(id=id).first()
+
+def get_workout_by_name(name):
+    return Workout.query.filter_by(name=name).first()
