+++ App/models/__init__.py
@@ -1,1 +1,3 @@
 from .user import *
+from .routine import *
+from .workout import *
