+++ App/templates/index.html
@@ -80,7 +80,8 @@
             </div>
 
             <div class="card-actions">
-                <form method = "POST" action = "/routine/create">
+                <form method = "POST" action = "">
+                    <!-- /routine/create -->
                     <label for="name_inline">Add</label>
                     <input class="btn waves-effect waves-light center" type="submit" value="submit">
                 </form>
