+++ App/templates/layout.html
@@ -19,9 +19,7 @@
               <a href="#!" class="brand-logo center">{% block page %}{% endblock %}</a>
               <ul id="nav-mobile" class="left">
                   <li><a href="/users">User Name</a></li>
-
                     <li><a href="/identify">Identify</a></li>
-                  {% endif %}
                   <li><a href="/static/users">Users JS</a></li> -->
               </ul>
               {% if is_authenticated %}
